backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
langchain==0.0.27